[#43] dodanie wszedzie CamelCase
