+++ sql_app/schemas_package/paczka_danych_schemas.py
@@ -2,19 +2,9 @@ from typing import List, Optional
 
 from pydantic import BaseModel, Field
 
+from sql_app.schemas_package.ogolne_schemas import CamelSchema
 from sql_app.schemas_package.wartosc_pomiaru_sensora_schemas import WartoscPomiaruSensoraSchema, \
     WartoscPomiaruProstaSchema
-
-
-def to_camel(string):
-    return ''.join(word.capitalize() for word in string.split('_'))
-
-
-class CamelSchema(BaseModel):
-    class Config:
-        alias_generator = to_camel
-        allow_population_by_field_name = True
-        orm_mode = True
 
 
 #czas updatuj po prostu na aktualny czas
